+++ crates/e2e/macro/src/config.rs
@@ -22,9 +22,9 @@ pub enum Backend {
 
     /// The lightweight approach skipping node layer.
     ///
-    /// This runs a runtime emulator within `TestExternalities` (using drink! library) in
+    /// This runs a runtime emulator within `TestExternalities`
     /// the same process as the test.
-    #[cfg(any(test, feature = "drink"))]
+    #[cfg(any(test, feature = "sandbox"))]
     RuntimeOnly(RuntimeOnly),
 }
 
@@ -61,9 +61,8 @@ impl Node {
     }
 }
 
-/// The runtime emulator that should be used within `TestExternalities` (using drink!
-/// library).
-#[cfg(any(test, feature = "drink"))]
+/// The runtime emulator that should be used within `TestExternalities`
+#[cfg(any(test, feature = "sandbox"))]
 #[derive(Clone, Eq, PartialEq, Debug, darling::FromMeta)]
 pub enum RuntimeOnly {
     #[darling(word)]
@@ -72,11 +71,11 @@ pub enum RuntimeOnly {
     Sandbox(syn::Path),
 }
 
-#[cfg(any(test, feature = "drink"))]
+#[cfg(any(test, feature = "sandbox"))]
 impl From<RuntimeOnly> for syn::Path {
     fn from(value: RuntimeOnly) -> Self {
         match value {
-            RuntimeOnly::Default => syn::parse_quote! { ::ink_e2e::MinimalSandbox },
+            RuntimeOnly::Default => syn::parse_quote! { ::ink_e2e::DefaultSandbox },
             RuntimeOnly::Sandbox(path) => path,
         }
     }
@@ -151,7 +150,7 @@ mod tests {
     #[test]
     fn config_works_runtime_only_with_custom_backend() {
         let input = quote! {
-            backend(runtime_only(sandbox = ::ink_e2e::MinimalSandbox)),
+            backend(runtime_only(sandbox = ::ink_e2e::DefaultSandbox)),
         };
         let config =
             E2EConfig::from_list(&NestedMeta::parse_meta_list(input).unwrap()).unwrap();
@@ -159,7 +158,7 @@ mod tests {
         assert_eq!(
             config.backend(),
             Backend::RuntimeOnly(RuntimeOnly::Sandbox(
-                syn::parse_quote! { ::ink_e2e::MinimalSandbox }
+                syn::parse_quote! { ::ink_e2e::DefaultSandbox }
             ))
         );
     }