+++ crates/e2e/sandbox/src/api.rs
@@ -1,0 +1,13 @@
+pub mod balance_api;
+pub mod contracts_api;
+pub mod system_api;
+pub mod timestamp_api;
+
+pub mod prelude {
+    pub use super::{
+        balance_api::BalanceAPI,
+        contracts_api::ContractAPI,
+        system_api::SystemAPI,
+        timestamp_api::TimestampAPI,
+    };
+}